updated hyperparameter locations in config
