+++ scripts/config/train_config.py
@@ -3,6 +3,8 @@ import logging
 import sys
 from pathlib import Path
 
+import torch.cuda
+
 # set random seed
 seed = 42
 
@@ -62,7 +64,7 @@ logging_strategy = 'epoch'
 load_best_model_at_end = True
 dataloader_pin_memory = False
 metric_for_best_model = 'f1'
-use_cpu = True
+use_cpu = False if torch.cuda.is_available() else True
 warmup_steps = 500
 
 ######################################################
@@ -70,20 +72,30 @@ warmup_steps = 500
 ######################################################
 # these parameters will be used with the NER model
 # but NOT with the document-level classifier
-ner_num_epochs = 4  # 50
-ner_per_device_train_batch_size = 4  # 32
+# generic parameters
 ner_per_device_eval_batch_size = 64
 ner_logging_dir = f'{savepath}/ner/logs'
+ner_num_splits = 3
 ner_weight_decay = 0.001
-ner_num_splits = 5
+
+# parameters for hyperparameter tuning
+ner_num_epochs = [1, 2, 4]
+ner_per_device_train_batch_size = [1, 2, 4]
+ner_lr_min = 1e-6
+ner_lr_max = 1e-4
 
 ######################################################
 ############# CLS MODEL PARAMETERS ###################
 ######################################################
 # these parameters will be used with the document
 # classifier but NOT with the NER model
-cls_num_epochs = 4  # 50
-cls_per_device_train_batch_size = 4  # 32
+# generic parameters
 cls_per_device_eval_batch_size = 64
 cls_logging_dir = f'{savepath}/classifier/logs'
 cls_weight_decay = 0.001
+
+# parameters for hyperparameter tuning
+cls_num_epochs = [1, 2, 4]  # 50
+cls_per_device_train_batch_size = [1, 2, 4]  # 32
+cls_lr_min = 1e-6
+cls_lr_max = 1e-4
