+++ src/mavpdl1/utils/utils.py
@@ -82,6 +82,7 @@ def optuna_hp_space(trial):
     """
     Get a hyperparameter space for optuna backend to use with a
     trainer trial
+    ONLY used with classifier currently
     :param trial: a trainer trial
     :return:
     """
@@ -90,6 +91,41 @@ def optuna_hp_space(trial):
         "per_device_train_batch_size": trial.suggest_categorical("per_device_train_batch_size", [1, 2, 4]),
         "num_train_epochs": trial.suggest_categorical("num_train_epochs", [1, 2, 4])
     }
+
+
+def calc_best_hyperparams(list_of_param_dicts):
+    """
+    Using the best hyperparameters of each inner loop,
+    as determined by outer loop holdout performance,
+    find best parameters through averaging
+    :param list_of_param_dicts:
+    :return:
+    """
+    # names of hyperparameters that must be ints
+    # todo: add others as needed
+    logging.info("There are currently only two hyperparameters listed"
+                 "as needing to be ints. If you include a hyperparameter"
+                 "in the search that needs to be an int and is NOT "
+                 "number of train epochs or per device train batch size, "
+                 "please add this to 'int_params' in function"
+                 "calc_best_hyperparams")
+    int_params = ["num_train_epochs", "per_device_train_batch_size"]
+    # holder for best parameters
+    best_params = {}
+    for param_dict in list_of_param_dicts:
+        for param, p_val in param_dict.items():
+            if param not in best_params:
+                best_params[param] = [p_val]
+            else:
+                best_params[param].append(p_val)
+    for param in best_params.keys():
+        if param in int_params:
+            p_val = round(sum(best_params[param]) / float(len(best_params[param])))
+        else:
+            p_val = sum(best_params[param]) / float(len(best_params[param]))
+        best_params[param] = p_val
+
+    return best_params
 
 
 def condense_df(df, label_encoder, gold_types="both"):
